refactor: Move ChunkLifecycleAction to the data_types crate
