+++ data_types/src/chunk_metadata.rs
@@ -62,6 +62,35 @@ impl ChunkStorage {
     }
 }
 
+/// Any lifecycle action currently in progress for this chunk
+#[derive(Debug, Copy, Clone, PartialEq)]
+pub enum ChunkLifecycleAction {
+    /// Chunk is in the process of being moved to the read buffer
+    Moving,
+
+    /// Chunk is in the process of being written to object storage
+    Persisting,
+
+    /// Chunk is in the process of being compacted
+    Compacting,
+}
+
+impl std::fmt::Display for ChunkLifecycleAction {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.name())
+    }
+}
+
+impl ChunkLifecycleAction {
+    pub fn name(&self) -> &'static str {
+        match self {
+            Self::Moving => "Moving to the Read Buffer",
+            Self::Persisting => "Persisting to Object Storage",
+            Self::Compacting => "Compacting",
+        }
+    }
+}
+
 /// Represents metadata about the physical storage of a chunk in a
 /// database.
 #[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Clone, Serialize, Deserialize)]
