+++ server/src/db/catalog/chunk.rs
@@ -3,13 +3,12 @@ use std::sync::Arc;
 use chrono::{DateTime, Utc};
 use snafu::Snafu;
 
-use data_types::chunk_metadata::ChunkAddr;
+use data_types::chunk_metadata::{ChunkAddr, ChunkLifecycleAction};
 use data_types::{
     chunk_metadata::{ChunkColumnSummary, ChunkStorage, ChunkSummary, DetailedChunkSummary},
     partition_metadata::TableSummary,
 };
 use internal_types::schema::Schema;
-use lifecycle::ChunkLifecycleAction;
 use metrics::{Counter, Histogram, KeyValue};
 use mutable_buffer::chunk::{snapshot::ChunkSnapshot as MBChunkSnapshot, MBChunk};
 use parquet_file::chunk::ParquetChunk;
